Improved DrpgCommandRequest.respond, using new DrpgResponse type
- Follow up interactions if they have already been replied to
- Removed `DrpgCommandResponse` class
- Replaced with `RichEmbed`
- Respond now accepts either any `DrpgResponse`, which is a `EmbedBuilder`, `EmbedBuilder[]`, `BaseMessageOptions`, or `String`

The new `RichEmbed` will return a `BaseMessageOptions` - which is the type shared by both Interactions and Messages when calling `.reply` on those types.

---

Going forward, it is safe to use `DrpgResponse` anywhere that typically would have returned an `EmbedBuilder`, and using this, you can use `respond` on any `DrpgCommandRequest`, regardless of the type.