Updated ping command example
The ping command has now been updated following the `DrpgResponse` improvement made in #6

the ping command is now a good example to follow if you want to create 'Rich Commands' - commands that support both slash and message commands.

These types of commands are best of gathering *only* the required parameters (parsing the `args`) from the message or the interaction - and then passing off into a common, shared implementation - meaning the logic is only required once, and it will support both.

It's now a clean and easily readable example that includes
- `DrpgResponse`
- Configuration of a Slash Command (`ChatInputcommand`)
- configuration of a prefix command (`messageRun`)
- Both entry points share the same output and processing